add: break FR into modules. Add FRUtil helper class. wip - make  ZoomCapture a stateless componenet
